+++ firebase-firestore/src/main/java/com/google/firebase/firestore/model/mutation/MutationBatch.java
@@ -17,6 +17,7 @@ package com.google.firebase.firestore.model.mutation;
 import static com.google.firebase.firestore.util.Assert.hardAssert;
 
 import com.google.firebase.Timestamp;
+import com.google.firebase.database.collection.ImmutableSortedMap;
 import com.google.firebase.firestore.model.DocumentKey;
 import com.google.firebase.firestore.model.MaybeDocument;
 import java.util.HashSet;
@@ -39,14 +40,34 @@ public final class MutationBatch {
    */
   public static final int UNKNOWN = -1;
 
+  /** The unique ID of this mutation batch. */
   private final int batchId;
+
+  /** The original write time of this mutation. */
   private final Timestamp localWriteTime;
+
+  /**
+   * Mutations that are used to populate the base values when this mutation is applied locally. This
+   * can be used to locally overwrite values that are persisted in the remote document cache. Base
+   * mutations are never sent to the backend.
+   */
+  private final List<Mutation> baseMutations;
+
+  /**
+   * The user-provided mutations in this mutation batch. User-provided mutations are applied both
+   * locally and remotely on the backend.
+   */
   private final List<Mutation> mutations;
 
-  public MutationBatch(int batchId, Timestamp localWriteTime, List<Mutation> mutations) {
+  public MutationBatch(
+      int batchId,
+      Timestamp localWriteTime,
+      List<Mutation> baseMutations,
+      List<Mutation> mutations) {
     hardAssert(!mutations.isEmpty(), "Cannot create an empty mutation batch");
     this.batchId = batchId;
     this.localWriteTime = localWriteTime;
+    this.baseMutations = baseMutations;
     this.mutations = mutations;
   }
 
@@ -98,8 +119,18 @@ public final class MutationBatch {
           maybeDoc.getKey());
     }
 
+    // First, apply the base state. This allows us to apply non-idempotent transform against a
+    // consistent set of values.
+    for (int i = 0; i < baseMutations.size(); i++) {
+      Mutation mutation = baseMutations.get(i);
+      if (mutation.getKey().equals(documentKey)) {
+        maybeDoc = mutation.applyToLocalView(maybeDoc, maybeDoc, localWriteTime);
+      }
+    }
+
     MaybeDocument baseDoc = maybeDoc;
 
+    // Second, apply all user-provided mutations.
     for (int i = 0; i < mutations.size(); i++) {
       Mutation mutation = mutations.get(i);
       if (mutation.getKey().equals(documentKey)) {
@@ -107,6 +138,23 @@ public final class MutationBatch {
       }
     }
     return maybeDoc;
+  }
+
+  /** Computes the local view for all provided documents given the mutations in this batch. */
+  public ImmutableSortedMap<DocumentKey, MaybeDocument> applyToLocalDocumentSet(
+      ImmutableSortedMap<DocumentKey, MaybeDocument> maybeDocumentMap) {
+    // TODO(mrschmidt): This implementation is O(n^2). If we iterate through the mutations first
+    // (as done in `applyToLocalView(DocumentKey k, MaybeDoc d)`), we can reduce the complexity to
+    // O(n).
+
+    ImmutableSortedMap<DocumentKey, MaybeDocument> mutatedDocuments = maybeDocumentMap;
+    for (DocumentKey key : getKeys()) {
+      MaybeDocument mutatedDocument = applyToLocalView(key, mutatedDocuments.get(key));
+      if (mutatedDocument != null) {
+        mutatedDocuments = mutatedDocuments.insert(mutatedDocument.getKey(), mutatedDocument);
+      }
+    }
+    return mutatedDocuments;
   }
 
   @Override
@@ -121,6 +169,7 @@ public final class MutationBatch {
     MutationBatch that = (MutationBatch) o;
     return batchId == that.batchId
         && localWriteTime.equals(that.localWriteTime)
+        && baseMutations.equals(that.baseMutations)
         && mutations.equals(that.mutations);
   }
 
@@ -128,6 +177,7 @@ public final class MutationBatch {
   public int hashCode() {
     int result = batchId;
     result = 31 * result + localWriteTime.hashCode();
+    result = 31 * result + baseMutations.hashCode();
     result = 31 * result + mutations.hashCode();
     return result;
   }
@@ -138,6 +188,8 @@ public final class MutationBatch {
         + batchId
         + ", localWriteTime="
         + localWriteTime
+        + ", baseMutations="
+        + baseMutations
         + ", mutations="
         + mutations
         + ')';
@@ -164,7 +216,16 @@ public final class MutationBatch {
     return localWriteTime;
   }
 
+  /** @return The user-provided mutations in this mutation batch. */
   public List<Mutation> getMutations() {
     return mutations;
+  }
+
+  /**
+   * @return The mutations that are used to populate the base values when this mutation batch is
+   *     applied locally.
+   */
+  public List<Mutation> getBaseMutations() {
+    return baseMutations;
   }
 }