+++ firebase-firestore/CHANGELOG.md
@@ -1,9 +1,12 @@
 # Unreleased
-- [feature] You can now query across all collections in your database with a
-  given collection ID using the `FirebaseFirestore.collectionGroup()` method.
+- [feature] Added `FieldValue.increment()`, which can be used in `update()`
+  and `set(..., SetOptions.merge())` to increment or decrement numeric field
+  values safely without transactions.
 - [fixed] Fixed calculation of SQLite database size on Android 9 Pie devices.
   Previous method could be off by a few MBs on these devices, potentially
   delaying garbage collection.
+- [feature] You can now query across all collections in your database with a
+  given collection ID using the `FirebaseFirestore.collectionGroup()` method.
 
 # 18.0.1
 - [fixed] Fixed an issue where Firestore would crash if handling write batches